Improve ZeroJointCommand and add a zeroJointCommand to JointSubsystem.
